+++ packages/util-crypto/src/nacl/decrypt.ts
@@ -1,7 +1,7 @@
 // Copyright 2017-2023 @polkadot/util-crypto authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import nacl from 'tweetnacl';
+import { naclSecretboxOpen } from './tweetnacl.js';
 
 /**
  * @name naclDecrypt
@@ -18,5 +18,5 @@ import nacl from 'tweetnacl';
  * ```
  */
 export function naclDecrypt (encrypted: Uint8Array, nonce: Uint8Array, secret: Uint8Array): Uint8Array | null {
-  return nacl.secretbox.open(encrypted, nonce, secret) || null;
+  return naclSecretboxOpen(encrypted, nonce, secret);
 }