+++ packages/util-crypto/src/random/asU8a.ts
@@ -27,4 +27,4 @@ export function randomAsU8a (length = 32): Uint8Array {
  * @name randomAsHex
  * @description Creates a hex string filled with random bytes.
  */
-export const randomAsHex = createAsHex(randomAsU8a);
+export const randomAsHex = /*#__PURE__*/ createAsHex(randomAsU8a);