+++ packages/util-crypto/src/sr25519/deriveHard.ts
@@ -5,4 +5,4 @@ import { sr25519DeriveKeypairHard } from '@polkadot/wasm-crypto';
 
 import { createDeriveFn } from './derive.js';
 
-export const sr25519DeriveHard = createDeriveFn(sr25519DeriveKeypairHard);
+export const sr25519DeriveHard = /*#__PURE__*/ createDeriveFn(sr25519DeriveKeypairHard);