+++ packages/x-bigint/src/mod.ts
@@ -1,8 +1,4 @@
 // Copyright 2017-2023 @polkadot/x-bigint authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export { packageInfo } from './packageInfo.js';
-
-const _BigInt = BigInt;
-
-export { _BigInt as BigInt };
+export * from './index.js';