+++ packages/x-randomvalues/src/fallback.ts
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Adapted from https://github.com/LinusU/react-native-get-random-values/blob/85f48393821c23b83b89a8177f56d3a81dc8b733/index.js
+//
 // Copyright (c) 2018, 2020 Linus Unnebäck
 // SPDX-License-Identifier: MIT